Feature/ws windows (#86)
* try to build ws on window on travis :postbox:

* cmake invocation fixed on windows :honeybee:

* Can use mbedtls to calculate hmac + no openssl config option

* build only windows on travis :clock730:

* run msbuild :pound:

* proper command to build :clock12:

* some build fixes

* change weird sizeof call :octopus:

* warning and missing includes fixes :white_flower:

* ifdef out statsd code on windows :snake:

* logic invertion in ifdef :lips:

* bring back makefile :scroll:

* command line tool is built with mbedtls :hospital:

+++ ws/ws_send.cpp
@@ -94,7 +94,7 @@ namespace ix
         std::streamoff size = file.tellg();
         file.seekg(0, file.beg);
 
-        memblock.resize(size);
+        memblock.resize((size_t) size);
         file.read((char*)&memblock.front(), static_cast<std::streamsize>(size));
 
         return memblock;
